feat: Add API slice for fetching users and notes
Created a new API slice called usersApiSlice and notesApiSlice which defines the endpoints and provides the necessary selectors for accessing user and notes data in the Redux store.

Added the necessary selectors to the various apiSlice files.

Updated the Redux store configuration in [frontend/src/app/store.js] to include the usersApiSlice reducer.

Imported and used the necessary dependencies in the User component, such as useGetUsersQuery, useSelector, and selectUserById.

These changes improve the functionality of the application by allowing the display of user information in the UI and fetching of user data from the API.